Add files via upload
This project included working on a real life sales dataset, with information on customers, transactions, products and markets, using Structured Query Language (SQL) to gather an idea of the dataset and to verify the calculations done later during the data visualization using Power BI.

Analysed the data to generate key insights on revenue, profit, and Top N customers and products. Used SQL queries to cross check the values I received while working on the visualization on Power BI, to maintain data accuracy.

Created Power BI dashboards with interactive slicers to drill down the Revenue, Profit %, Profit contribution %, total revenue, total quantity sold, and total profit % by years and months, which provided a more concise idea of the revenue and profit trends across the years.

Tools used-
SQL: To load the dataset into MySQL and write queries to get an idea of the dataset I'm working on, and also to verify the calculations done on Power BI.

Power BI: To load the database, clean the data, build relationships between tables and visualize dashboards to provide key insights.